Update header link

+++ src/component/Decks/Deck/Deck.jsx
@@ -1,4 +1,9 @@
-import React from 'react';
+import React, { useMemo } from 'react';
+import { IconContext } from 'react-icons';
+import { BiEditAlt } from 'react-icons/bi';
+import { MdDeleteForever } from 'react-icons/md';
+import { AiOutlineFileAdd } from 'react-icons/ai';
+import { Tooltip as ReactTooltip } from 'react-tooltip';
 import PropTypes from 'prop-types';
 import { useNavigate } from 'react-router-dom';
 import axios from 'axios';
@@ -14,12 +19,18 @@ function Deck({
   deckImageThree,
 }) {
   const { setChoiceDeck } = useLogin();
+  const addIcon = useMemo(() => ({ className: 'global-class-add-card' }), []);
+  const editIcon = useMemo(() => ({ className: 'global-class-edit' }), []);
+  const deleteIcon = useMemo(() => ({ className: 'global-class-delete' }), []);
 
   const navigate = useNavigate();
 
   const handleDeleteDeck = async () => {
-    await axios.delete(`http://localhost:8000/api/deck/deckdelete/${idDeck}`);
-    getDeck();
+    const result = window.confirm('supprimer ce deck definitivement ?');
+    if (result) {
+      await axios.delete(`http://localhost:8000/api/deck/deckdelete/${idDeck}`);
+      getDeck();
+    }
   };
 
   const handleChoiceAddCards = () => {
@@ -33,21 +44,48 @@ function Deck({
 
   return (
     <div className="deck_base">
+      <div className="name_deck">
+        <h3>{deckName}</h3>
+      </div>
       <div className="card_deck">
-        <h4>{deckName}</h4>
         <img className="image1" src={deckImageOne} alt="" />
         <img src={deckImageTwo} alt="" />
         <img className="image3" src={deckImageThree} alt="" />
       </div>
-      <button type="button" onClick={handleChoiceAddCards}>
-        ajout carte
-      </button>
-      <button type="button" onClick={handleDeleteDeck}>
-        supprimer deck
-      </button>
-      <button type="button" onClick={handleChoiceModifDeck}>
-        modif deck
-      </button>
+      <div className="logo">
+        <IconContext.Provider value={addIcon}>
+          <AiOutlineFileAdd
+            size={30}
+            id="icon-add-card"
+            onClick={handleChoiceAddCards}
+          />
+        </IconContext.Provider>
+        <IconContext.Provider value={editIcon}>
+          <BiEditAlt size={30} id="icon-edit" onClick={handleChoiceModifDeck} />
+        </IconContext.Provider>
+        <IconContext.Provider value={deleteIcon}>
+          <MdDeleteForever
+            size={30}
+            id="icon-delete"
+            onClick={handleDeleteDeck}
+          />
+        </IconContext.Provider>
+      </div>
+      <ReactTooltip
+        anchorId="icon-add-card"
+        place="bottom"
+        content="Ajouter les cartes dans le deck"
+      />
+      <ReactTooltip
+        anchorId="icon-edit"
+        place="bottom"
+        content="Modifier image et nom du deck"
+      />
+      <ReactTooltip
+        anchorId="icon-delete"
+        place="bottom"
+        content="Supprimer le deck"
+      />
     </div>
   );
 }
